# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: C7NTAX API
  description: |
    Professional Services Automation API — programmatic access to the C7NTAX platform.
    All endpoints require Bearer token authentication (JWT) obtained via `/api/auth/login`.
    External apps and user-created plugins use this same API.

    ## Authentication
    1. POST `/api/auth/login` with email + password
    2. If MFA is enabled, receive `mfaToken` and POST `/api/auth/verify-mfa` with TOTP code
    3. Use returned `token` as `Authorization: Bearer <token>` header

    ## Rate Limiting
    100 requests per 15-minute window per token. 429 responses include `Retry-After`.

    ## Permissions
    Each endpoint lists required permission(s). Client users are scoped to their company's data.
  version: "1.0.0"
  contact:
    name: Cyber 7 Group
    email: support@cyber7group.com
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.C7NTAX.com/api
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /tickets:
    get:
      tags: [Tickets]
      summary: List tickets with filtering
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/TicketStatus" } }
        - { name: priority, in: query, schema: { $ref: "#/components/schemas/TicketPriority" } }
        - { name: boardId, in: query, schema: { type: string } }
        - { name: companyId, in: query, schema: { type: string } }
        - { name: assignedToId, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Paginated ticket list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Ticket" } }
                  total: { type: integer }
    post:
      tags: [Tickets]
      summary: Create a new ticket
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, boardId]
              properties:
                title: { type: string }
                description: { type: string }
                boardId: { type: string }
                companyId: { type: string }
                priority: { $ref: "#/components/schemas/TicketPriority" }
                source: { $ref: "#/components/schemas/TicketSource" }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/Ticket" } } } }
  /tickets/{id}:
    get:
      tags: [Tickets]
      summary: Get ticket details with notes and time entries
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/TicketDetail" } } } }
    patch:
      tags: [Tickets]
      summary: Update ticket status, assignment, or fields
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                status: { $ref: "#/components/schemas/TicketStatus" }
                priority: { $ref: "#/components/schemas/TicketPriority" }
                boardId: { type: string }
                assignedToId: { type: string }
                dueDate: { type: string, format: date-time }
                estimatedHours: { type: number }
                note: { type: string, description: "Note to add with this update" }
                noteInternal: { type: boolean, description: "Mark note as internal only" }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Ticket" } } } }
  /tickets/{id}/notes:
    post:
      tags: [Tickets]
      summary: Add a note to a ticket
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                isInternal: { type: boolean, default: false }
      responses:
        "201": { description: "Note added" }
  /tickets/{id}/time:
    post:
      tags: [Tickets]
      summary: Log time entry against a ticket
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startTime, endTime]
              properties:
                startTime: { type: string, format: date-time }
                endTime: { type: string, format: date-time }
                description: { type: string }
                billable: { type: boolean, default: true }
      responses:
        "201": { description: "Time entry logged" }

  # ─── Boards ───
...
  /integrations/types:
    get:
      tags: [Integrations]
      summary: List available integration service types
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
                example: ["flexpoint", "quickbooks", "pax8", "avanan", "proofpoint", "sentinelone", "itglue", "microsoft365", "azure", "aws"]
  /integrations:
    get:
      tags: [Integrations]
      summary: List configured integrations
      security: [{ bearerAuth: [] }]
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/IntegrationConfig" } } } } }
    post:
      tags: [Integrations]
      summary: Add a new integration configuration
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, name, credentials]
              properties:
                kind: { type: string }
                name: { type: string }
                credentials: { type: object }
                settings: { type: object }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/IntegrationConfig" } } } }
  /integrations/{id}/test:
    post:
      tags: [Integrations]
      summary: Test integration connection
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
  /integrations/{id}/sync:
    post:
      tags: [Integrations]
      summary: Trigger manual sync for an integration
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  kind: { type: string }
                  recordsProcessed: { type: integer }
                  errors: { type: array, items: { type: string } }
                  syncedAt: { type: string, format: date-time }

  # ─── WebSocket ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SystemRole:
      type: string
      enum: [admin, manager, technician, dispatcher, billing_manager, client_admin, client_user, read_only]
    TicketStatus:
      type: string
      enum: [new, in_progress, waiting_on_client, waiting_on_third_party, on_hold, pending_approval, resolved, closed, cancelled]
    TicketPriority:
      type: string
      enum: [low, medium, high, critical]
    TicketSource:
      type: string
      enum: [phone, email, portal, chat, monitoring, walk_in, api, internal]
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        role: { $ref: "#/components/schemas/SystemRole" }
        status: { type: string }
        mfaEnabled: { type: boolean }
        companyId: { type: string }
        lastLoginAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    Ticket:
      type: object
      properties:
        id: { type: string }
        ticketNumber: { type: string, example: "C7-MK1234-AB12" }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/TicketStatus" }
        priority: { $ref: "#/components/schemas/TicketPriority" }
        source: { $ref: "#/components/schemas/TicketSource" }
        boardId: { type: string }
        companyId: { type: string }
        assignedToId: { type: string }
        createdById: { type: string }
        waitingSince: { type: string, format: date-time }
        followUpCount: { type: integer }
        dueDate: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TicketDetail:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            company: { $ref: "#/components/schemas/Company" }
            assignedTo: { $ref: "#/components/schemas/User" }
            board: { $ref: "#/components/schemas/Board" }
            notes: { type: array, items: { type: object } }
            timeEntries: { type: array, items: { type: object } }
    Board:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        autoCloseDays: { type: integer }
        createdAt: { type: string, format: date-time }
...
    Company:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        zip: { type: string }
        country: { type: string }
        website: { type: string }
        notes: { type: string }
        status: { type: string }
...
    IntegrationConfig:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        status: { type: string, enum: [connected, error, disconnected] }
        lastSyncAt: { type: string, format: date-time }
        errorMessage: { type: string }